requirements.txt
uvicorn~=0.13.3
fastapi~=0.63.0
tortoise-orm~=0.16.21